Support Debian (and other) base images (#39)
* Calculate image partition offsets rather than hard-coded values
Start generalizing base image handling

* Fix copy/paste error in partition offset calculation
Add fdisk dependency to Dockerfile

* Troubleshooting partition block determination

* Troubleshoot mount points in fdisk output

* Update sj201 setup for Debian compat
Read base image from envvar and update docs

* Override resolv.conf for debian base image compat

* Add debian boot overlay files
Update default core branch ref
Cleanup logging
Update boot overlay handling to support multiple images
Add build directory to gitignore
Extract default username and image name to variables
Troubleshoot sj201 installation dependencies

* Fix typo in distribution check
Skip linux header installation under Ubuntu base

* Make Debian patches specific to Debian base images
Fix kernel check for backwards-compat

* Troubleshooting kernel selection

* Fix typos in module build dir detection

* Add root and profile.d overlay files
Abstract password to variable
Fix typos
Replace dnsmasq with systemd-resolved to fix wifi-connect
Update automation to better mount image

* Fix typos in mount commands

* Add default bashrc overlay
Cleanup debian dependencies moved to godebos base image setup

* Cleanup debian dependencies moved to godebos base image setup
Disable dnsmasq service for wifi-connect compat.

* Minor wifi-setup service and setup fixes

* Cleanup image unmounting
Add losetup dependency

* Cleanup Debian image changes

* Update scripts for Debian base image compat.

* Fix mount dependency in Dockerfile
Update automation to handle losetup mount/unmount of image

* Troubleshoot Docker build errors

* Specify loop device in prepare.sh instead of using next available

* Fix bugs in updated loop changes

* Update docker example to include required `/dev` mount to support loop device

* Troubleshoot docker partition loop devices

* Troubleshoot loop device setup

* Fix error removing loop devices

* Hard-code loop device used in losetup

* Update scripts to handle mounted image partitions

* Update dockerfile and boot overlay for debian base image

* Fix GUI patch for Ubuntu builds

* Troubleshoot image full in Core installation
Fix missed hard-coded img reference in build scripts

* Remove package caches

* Fix default root .bashrc
Cleanup deprecated code
Ensure GUI starts before skills and enclosure services
Update default docker example in docs

* Update default config
Make Neon wait for gui service start
Deprecate Ubuntu attempt to build keyboard since dependencies are incompatible
Simplify prepare.sh

* Update setup skill settings

* pip install homescreen skill

* Update default config overlay

* Handle runtime image overlay files

* Update skill settings overlay
Troubleshoot dashboard dependencies
Catch errors in cleanup automation
Update README.md

* Update setup skill overlay
Troubleshoot full image during build

* Update default config
Add pishrink resizing before compressing output image

* Add pishrink deps to Dockerfile
Catch and fix issue with missing skel files

* Troubleshoot skel dir contents

* Troubleshoot skel dir contents

* Fix skel dir overlay patch

* Update default logging and user settings skill configs

* Update system config to remove default value overrides

* Fix default config path in overlay

* Relocate swapfile during build to resolve FS full errors

* Add logging around swapfile handling
Handle pishrink error if image already minimized

* Add default user to `audio` group for MPRIS

+++ automation/cleanup.sh
@@ -31,11 +31,35 @@ build_dir=${1}
 cd "${build_dir}" || exit 10
 
 # Re-enable restart prompts
-sudo mv mnt/etc/apt/apt.conf.d/.99needrestart mnt/etc/apt/apt.conf.d/99needrestart
-sudo mv mnt/root/bashrc mnt/root/.bashrc
+if [ -f mnt/etc/apt/apt.conf.d/.99needrestart ]; then
+    sudo mv mnt/etc/apt/apt.conf.d/.99needrestart mnt/etc/apt/apt.conf.d/99needrestart
+fi
+
+# Cleanup overridden resolv.conf
+sudo rm mnt/etc/resolv.conf
+if [ -f mnt/etc/.resolv.conf ]; then
+    sudo mv mnt/etc/.resolv.conf mnt/etc/resolv.conf
+fi
+
+# Cleanup root bashrc
+sudo rm mnt/root/.bashrc
+if [ -f mnt/root/bashrc ]; then
+    echo "Restoring root .bashrc file"
+    sudo mv mnt/root/bashrc mnt/root/.bashrc
+fi
+
+# Replace swapfile for build
+if [ -f "${build_dir}/swapfile" ]; then
+    echo "replacing swapfile"
+    sudo mv "${build_dir}/swapfile" mnt/swapfile
+fi
+
+#sudo mv mnt/root/bashrc mnt/root/.bashrc
 sudo rm -rf mnt/tmp/*
 echo "Temporary files removed"
+sudo umount mnt/boot/firmware || echo "boot partition not mounted"
 sudo umount mnt/run/systemd/resolve || exit 10
 sudo umount mnt || exit 10
+rm -r mnt
+sudo losetup -d /dev/loop99
 echo "Image unmounted"
-#rm -r mnt
